Slice 17: Split dashboardConfig into widgets.js, layouts.js, and a barrel
- Move WIDGETS, WIDGET_MAP, WIDGET_CATEGORIES, DEFAULT_VISIBLE into
  features/dashboard/config/widgets.js with file/JSDoc comments.
- Move autoLayout, buildResponsiveLayouts, LG_DEFAULT, DEFAULT_LAYOUTS,
  and LAYOUT_PRESETS into features/dashboard/config/layouts.js.
- Add features/dashboard/config/index.js barrel; consumers now import
  from "../config" instead of "../dashboardConfig".
- Update useDashboardLayout, WidgetCatalog, and DashboardGrid to use
  the new path; delete the old monolithic dashboardConfig.js.
- Update the ESLint max-lines override to point at config/widgets.js
  (the only remaining declarative-data file).
